ref: made some widgets shared
